secrets added

+++ srcs/requirements/wordpress/tools/wordpress.sh
@@ -1,36 +1,47 @@
 #!/bin/sh
 
+MYSQL_PASSWORD=$(cat /run/secrets/db_password.txt)
+WP_ADMIN_PASSWORD=$(cat /run/secrets/wp_admin_password.txt)
+WP_USER_PASSWORD=$(cat /run/secrets/wp_password.txt)
+
 mkdir -p /var/www/wordpress
 cd /var/www/wordpress
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+if [ ! -f /usr/local/bin/wp ]; then
+	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+	chmod +x wp-cli.phar
+	mv wp-cli.phar /usr/local/bin/wp
+fi
 
-wp core download --allow-root
+if [ ! -f wp-config.php ]; then
+	wp core download --allow-root
 
-cp /var/www/wordpress/wp-config-sample.php /var/www/wordpress/wp-config.php
+	cp /var/www/wordpress/wp-config-sample.php /var/www/wordpress/wp-config.php
 
-sed -i "s/define( 'DB_NAME', 'database_name_here' );/define( 'DB_NAME', '${MYSQL_DATABASE}' );/" wp-config.php
-sed -i "s/define( 'DB_USER', 'username_here' );/define( 'DB_USER', '${MYSQL_USER}' );/" wp-config.php
-sed -i "s/define( 'DB_PASSWORD', 'password_here' );/define( 'DB_PASSWORD', '${MYSQL_PASSWORD}' );/" wp-config.php
-sed -i "s/localhost/mariadb/" wp-config.php
+	sed -i "s/define( 'DB_NAME', 'database_name_here' );/define( 'DB_NAME', '${MYSQL_DATABASE}' );/" wp-config.php
+	sed -i "s/define( 'DB_USER', 'username_here' );/define( 'DB_USER', '${MYSQL_USER}' );/" wp-config.php
+	sed -i "s/define( 'DB_PASSWORD', 'password_here' );/define( 'DB_PASSWORD', '${MYSQL_PASSWORD}' );/" wp-config.php
+	sed -i "s/localhost/mariadb/" wp-config.php
+fi
 
+echo "Waiting for database connection..."
 sleep 10
 
-wp core install \
-	--url="${DOMAIN_NAME}" \
-	--title="${SITE_TITLE}" \
-	--admin_user="${WP_ADMIN_USER}" \
-	--admin_password="${WP_ADMIN_PASSWORD}" \
-	--admin_email="${WP_ADMIN_EMAIL}" \
-	--allow-root
+if ! wp core is-installed --allow-root; then
+	wp core install \
+		--url="${DOMAIN_NAME}" \
+		--title="${SITE_TITLE}" \
+		--admin_user="${WP_ADMIN_USER}" \
+		--admin_password="${WP_ADMIN_PASSWORD}" \
+		--admin_email="${WP_ADMIN_EMAIL}" \
+		--allow-root
 
-wp user create \
-	"${WP_USER}" "${WP_USER_EMAIL}" \
-	--user_pass="${WP_USER_PASSWORD}" \
-	--role=editor \
-	--allow-root
+	wp user create \
+		"${WP_USER}" "${WP_USER_EMAIL}" \
+		--user_pass="${WP_USER_PASSWORD}" \
+		--role=editor \
+		--allow-root
+fi
 
 if ! wp theme is-installed twentytwentyfour --allow-root; then
 		wp theme install twentytwentyfour --activate --allow-root
@@ -40,4 +51,7 @@ fi
 
 mkdir -p /run/php
 
-php-fpm7.4 -F
+PHP_VERSION=$(php -r 'echo PHP_MAJOR_VERSION.".".PHP_MINOR_VERSION;')
+
+echo "Starting PHP-FPM..."
+php-fpm${PHP_VERSION} -F
